fix: init Roadiz page data at layout level
