+++ layouts/default.vue
@@ -1,4 +1,36 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import type { RoadizNodesSources } from '@roadiz/types'
+
+// init Roadiz page data (i.e. dynamic page)
+callOnce(async () => {
+    const route = useRoute()
+
+    // check if it's a dynamic page
+    if (route.name !== 'slug') return
+
+    const { webResponse, alternateLinks } = await useRoadizWebResponse()
+
+    // fill page data
+    usePage({ webResponse, alternateLinks })
+
+    // init I18N
+    const nuxtApp = useNuxtApp()
+    await callOnce(async () => {
+        const locale = (webResponse?.item as RoadizNodesSources)?.translation?.locale
+
+        if (locale) {
+            // set the current global locale
+            await nuxtApp.$i18n.setLocale(locale)
+        }
+        else {
+            // get the locale from the route (prefix) or cookie ?
+        }
+
+        // init alternate links
+        useAlternateLinks(alternateLinks)
+    })
+})
+</script>
 
 <template>
     <div>
